added choose password screen with anim. placeholders
