+++ locales/en.json
@@ -32,9 +32,10 @@
 		"go_back": "Go Back",
 		"invalid_password": "Invalid password"
 	},
-	"create_wallet": {
-		"title": "Create New Wallet",
-		"password": "Password (min. 8 characters)",
+	"choose_password": {
+		"title": "Choose a password",
+		"subtitle": "You'll use this password to unlock your MetaMask wallet only on this device.",
+		"password": "New Password",
 		"confirm_password": "Confirm Password",
 		"create_button": "CREATE",
 		"import_with_seed_phrase": "Import with seed phrase",
